chore(code): Clean up CLI package
